Address GCC warnings
[ROCm/rocRoller commit: 5a34619a9ebb6f0b084c860fba0df0e5659dcd0d]

+++ shared/rocroller/lib/include/rocRoller/CodeGen/Arithmetic/ArithmeticBase_impl.hpp
@@ -71,7 +71,7 @@ namespace rocRoller
 
         dwords.clear();
         dwords.reserve(count);
-        for(int i = 0; i < count; i++)
+        for(size_t i = 0; i < count; i++)
         {
             dwords.push_back(input->subset({i}));
         }
